version 0.0.19

+++ setup.py
@@ -1,14 +1,17 @@
+import os
 import setuptools
 
-with open("version.txt", "r") as f:
-    version = f.read()
-    if version:
-        versionno = version.split("=")[-1].strip("' ")
-        arr = [int(item) for item in versionno.split(".")]
-        arr[2] += 1
-        version = "%d.%d.%d" % tuple(arr)
-        with open("version.txt", "w") as w:
-            w.write(version)
+version = "0.0.0"
+if os.path.isfile("version.txt"):
+    with open("version.txt", "r") as f:
+        version = f.read()
+        if version:
+            versionno = version.split("=")[-1].strip("' ")
+            arr = [int(item) for item in versionno.split(".")]
+            arr[2] += 1
+            version = "%d.%d.%d" % tuple(arr)
+            with open("version.txt", "w") as w:
+                w.write(version)
 
 setuptools.setup(
     name="gecosistema_database",
@@ -24,5 +27,6 @@ setuptools.setup(
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ),
-    install_requires=['gecosistema_core','unicodecsv']
+    #apt-get install unixodbc
+    install_requires=['gecosistema_core','unicodecsv','pyodbc']
 )
